Optimization of simplify 3D objects.
git-svn-id: svn://svn.code.sf.net/p/moka-modeller/code@82 578f8393-57f1-403d-81f3-a3b3dc990742

+++ src/lib-rounding/rounding-vertex.cc
@@ -704,6 +704,6 @@ void CRoundingVertex::setDartRoundingCoef(CDart* ADart, int ADimension,
 					  TCoordinate AValue) const
 {
   assert(ADimension == 0 || ADimension == 1);
-  ADart->setDirectInfo(FCoefs[ADimension], (void*) (int) (1000 * AValue));
+  ADart->setDirectInfo(FCoefs[ADimension], (void*) (long int) (1000 * AValue));
 }
 //******************************************************************************
